+++ src/commands/Sw.cc
@@ -3,8 +3,8 @@
 #include "InvalidScan.h"
 using namespace std;
 void Sw::run(MipsClient & client, const LabelTable & table) const {
-    int32_t s = client.getRegister(tokens.at(5).valAsNumber());
-    int32_t i = tokens.at(3).valAsNumber();
+    int64_t s = client.getRegister(tokens.at(5).valAsNumber());
+    int64_t i = tokens.at(3).valAsNumber();
     if ((s+i) % 4 != 0) throw InvalidScan("Unaligned Access to memory.");
     if (s + i > MAX_MEM * 4 || s + i < 0) throw InvalidScan("Memory access out of bounds.");
